Move TeamWaivers and WaiverPlayer to responses module
